Fix Int to Float when getting PhysicsSettings.Vertices.Radius in physics3.json parsing

+++ src/physics/cubismphysicsjson.ts
@@ -609,7 +609,7 @@ export namespace Live2DCubismFramework {
         .getValueByString(Vertices)
         .getValueByIndex(vertexIndex)
         .getValueByString(Radius)
-        .toInt();
+        .toFloat();
     }
 
     /**
